ml_service/requirements.txt
pandas
pymongo
fastapi
uvicorn[standard]
scikit-learn
joblib
python-dotenv
certifi==2026.2.25
cffi==2.0.0
cryptography==46.0.5
numpy==2.4.3
pycparser==3.0
setuptools==82.0.1
wheel==0.46.3
